Conflicts resolved

+++ payment-concentrator/src/main/java/com/sep/payment/paymentconcentrator/repository/PaymentRequestRepository.java
@@ -4,5 +4,5 @@ import com.sep.payment.paymentconcentrator.domain.entity.PaymentRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRequestRepository extends JpaRepository<PaymentRequest, Long> {
-    PaymentRequest findByMerchantOrderId(int merchantOrderId);
+    PaymentRequest findByMerchantOrderId(Long merchantOrderId);
 }
